# specs/001-idea-evolution-engine/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PoAppIdea API
  description: Self-Evolving Ideation Engine - API for idea generation, evolution, and artifact production
  version: 1.0.0
  contact:
    name: PoAppIdea Team
servers:
  - url: https://localhost:5001
    description: Local development (HTTPS)
  - url: http://localhost:5000
    description: Local development (HTTP)
  - url: https://poappidea.azurewebsites.net
    description: Production (Azure App Service)
...
paths:
...
  /api/sessions/{sessionId}:
    get:
      tags: [Session]
      summary: Get session details
      operationId: getSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{sessionId}/ideas:
    post:
      tags: [Spark]
      summary: Generate next batch of ideas
      description: Generates 10 ideas for the current batch, incorporating learning from previous swipes
      operationId: generateIdeas
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Batch of ideas generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdeaBatchResponse'
        '400':
          description: Session not in Phase 1 or batch limit reached
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{sessionId}/swipes:
    post:
      tags: [Spark]
      summary: Record a swipe on an idea
      operationId: recordSwipe
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordSwipeRequest'
      responses:
        '201':
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwipeResponse'
        '400':
          description: Invalid swipe data or idea already swiped
...
  /api/sessions/{sessionId}/visuals/{assetId}/select:
    post:
      tags: [Visual]
      summary: Select a visual direction
      operationId: selectVisual
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: assetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Visual selected
  /api/sessions/{sessionId}/artifacts:
    post:
      tags: [Artifacts]
      summary: Generate final artifacts
      description: Generates PRD, Technical Deep-Dive, and Visual Asset Pack
      operationId: generateArtifacts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '202':
          description: Artifact generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactGenerationStatus'

    get:
      tags: [Artifacts]
      summary: Get generated artifacts
      operationId: getArtifacts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtifactResponse'
...
  /api/gallery/{artifactId}/import:
    post:
      tags: [Gallery]
      summary: Import idea from gallery as session starting point
      operationId: importFromGallery
      security:
        - bearerAuth: []
      parameters:
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: New session created from import
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    AppType:
      type: string
      enum: [Game, Productivity, Mobile, Automation]
    SessionStatus:
      type: string
      enum: [InProgress, Completed, Abandoned]
    SessionPhase:
      type: string
      enum: [Phase0_Scope, Phase1_Spark, Phase2_Mutation, Phase3_FeatureExpansion, Submission, Phase4_ProductRefinement, Phase5_TechnicalRefinement, Phase6_Visual, Completed]
    SwipeDirection:
      type: string
      enum: [Like, Dislike]
...
    ArtifactType:
      type: string
      enum: [PRD, TechnicalDeepDive, VisualAssetPack]

    # Request/Response Schemas
...
    SessionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appType:
          $ref: '#/components/schemas/AppType'
        complexityLevel:
          type: integer
        currentPhase:
          $ref: '#/components/schemas/SessionPhase'
        status:
          $ref: '#/components/schemas/SessionStatus'
        createdAt:
          type: string
          format: date-time
...
    IdeaBatchResponse:
      type: object
      properties:
        batchNumber:
          type: integer
        ideas:
          type: array
          items:
            $ref: '#/components/schemas/IdeaResponse'
        remainingBatches:
          type: integer
    IdeaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        batchNumber:
          type: integer
    RecordSwipeRequest:
      type: object
      required: [ideaId, direction, durationMs]
      properties:
        ideaId:
          type: string
          format: uuid
        direction:
          $ref: '#/components/schemas/SwipeDirection'
        durationMs:
          type: integer
          minimum: 1
    SwipeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        speedCategory:
          type: string
          enum: [Fast, Medium, Slow]
...
    ArtifactGenerationStatus:
      type: object
      properties:
        status:
          type: string
          enum: [Queued, InProgress, Completed, Failed]
        artifacts:
          type: array
          items:
            type: string
            enum: [PRD, TechnicalDeepDive, VisualAssetPack]
    ArtifactResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/ArtifactType'
        title:
          type: string
        content:
          type: string
        blobUrl:
          type: string
          format: uri
        humanReadableSlug:
          type: string
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        traceId:
          type: string